Extract whitespace test classes.
